requirements.txt
#due_evaluator==0.0.8 ## You need to install this from github https://github.com/due-benchmark/evaluator
jsonlines==4.0.0
matplotlib==3.8.2
numpy==2.2.3
pandas==2.2.3
pdf2image==1.17.0
Pillow==10.1.0
Pillow==11.1.0
pymupdf==1.25.3
python_Levenshtein==0.27.1
scikit_learn==1.6.1
setuptools==65.5.0
torch==2.1.1
tqdm==4.66.5
transformers==4.35.0
wandb==0.19.8
